This creates a separate graph for each significant writer
This was made as it was possible that the initial graph made was so jagged due to the mixture of writers that breaking bad uses. In this you can see slightly more correlation in some cases, however sometimes this makes the correlation worse. In this code you are asked to pick which writer you want to graph. It is interesting however that each writer tends to have the same negative correlation.